# openapi (4).yaml
openapi: 3.0.1
info:
  title: Book Handler
  description: 'This is a BookHandler server. BookHandler is a platform for sharing and reading books and graphical novels that can be described as online library. This application should be especially useful for book enthusiasts that are looking for rare prints, not available in internet bookstores and on general auction sites and people interested in expanding their personal collections.'
  contact:
    email: 212829@student.pwr.edu.pl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
...
  /user:
    post:
      tags:
      - user
      summary: Create user
      description: Service's visitor can create User account 
      operationId: createUser
      requestBody:
        description: Creates user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        200:
          description: successful operation
          content: {}
        409:
          description: username already taken
          content: {}
      x-codegen-request-body-name: body
  /user/login:
    get:
      tags:
      - user
      summary: Log user into the system
      operationId: loginUser
      parameters:
      - name: username
        in: query
        description: The username for login
        required: true
        schema:
          type: string
      - name: password
        in: query
        description: The password for login in clear text
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
        400:
          description: Invalid username/password supplied
          content: {}
...
  /user/{username}:
    get:
      tags:
      - user
      summary: Get user by user name
      operationId: getUserByName
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        404:
          description: User not found
          content: {}
    put:
      tags:
      - user
      summary: Update user
      description: This can only be done by the admin or logged in owner of updated user account.
      operationId: updateUser
      parameters:
      - name: username
        in: path
        description: Name that needs to be updated
        required: true
        schema:
          type: string
      requestBody:
        description: Updated user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        404:
          description: User not found
          content: {}
        403:
          description: User does not have the rights to update this account
          content: {}
      x-codegen-request-body-name: body
    delete:
      tags:
      - user
      summary: Delete user
      description: This can only be done by the admin or logged in owner of user account.
      operationId: deleteUser
      parameters:
      - name: username
        in: path
        description: The name that needs to be deleted
        required: true
        schema:
          type: string
      - name: role
        in: cookie
        description: The user role
        required: true
        schema:
          type: string
          enum: 
            - admin
            - member
      responses:
        404:
          description: User not found
          content: {}
        403:
          description: User does not have the rights to delete account
          content: {}
...
components:
  schemas:
...
    User:
      required:
      - username
      - email
      - password
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        userRole:
          type: string
          description: User Status
          enum: 
            - admin
            - member
        dateAdded:
          type: string
          format: date-time
...